requirements.txt
Flask
langchain_community
langchain_core
langchain_ollama
python-dotenv
